requirements.txt
# ====== Python Dependencies ======
flask>=3.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
requests>=2.31.0
python-telegram-bot>=21.0
paho-mqtt>=2.0.0
redis>=5.0.0
qdrant-client>=1.7.0

# ====== Optional: Voice ======
# Porcupine wake word (requires access key)
# pvporcupine>=3.0.0
# sounddevice>=0.4.6
# pyaudio>=0.2.14
